Refactoring Api Route
