+++ server/routes/main.js
@@ -1,0 +1,7 @@
+const router = require("express").Router();
+
+router.get('/', (req,res)=>{
+    res.send('Welcome to Storz API v1.0!');
+})
+
+module.exports = router
